"Agregado el test del modelo Usuario"

+++ src/main/java/ar/edu/unq/criptop2p/model/entity/Intencion.java
@@ -57,7 +57,7 @@ public class Intencion {
         return getUsuario().getCantidadOperaciones();
     }
 
-    public byte getReputacion() {
+    public int getReputacion() {
         return getUsuario().getReputacion();
     }
